Use built-ins for hex conversion if available (node 25)
