+++ .gitignore
@@ -1,5 +1,5 @@
+.pnpm-store/
 coverage/
 docs/
 lib/
 node_modules/
-.pnpm-store/
